refactor(perms): keep scoped-permission DB ops under onyx/db
Address PR review on the scoped-manager primitives:
- move scoped_group_ids_subquery, the managed-group read, and
  within_managed_scope_clause into new onyx/db/scoped_permissions.py; the auth
  layer keeps only the bundle guard + gates and consumes that DB interface
- document get_scoped_groups(permission=None) scope-introspection path
- fix docs: the single-classifier model removed has_permission_or_scope, not
  has_permission
